requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
click>=8.1.0
gradio>=4.0.0
httpx>=0.25.0
langchain>=0.1.0
